third_party/WebKit/LayoutTests/inspector/sources/debugger/live-edit-no-reveal-expected.txt

Tests live edit feature.

Debugger was enabled.

Running: testLiveEditWhenPausedDoesNotCauseCursorMove
Script execution paused.
Moving cursor to (0, 0).
Committing live edit.
Script execution paused.
Script execution resumed.
Cursor position is: (0, 0).
Debugger was disabled.
